reorganise so it's pip installable
